doc-validator/requirements.txt
streamlit
openai
anthropic
requests
python-dotenv
PyPDF2
Pillow
pytesseract
scikit-learn
rank-bm25
packaging
numpy
pymilvus
finance-datareader
plotly
koreanize-matplotlib
finance_datareader
exa-search